Reorganize circle package.
