--- src/main/java/org/stellar/anchor/paymentsservice/circle/CircleWalletResponse.java
@@ -1,8 +1,0 @@
-package org.stellar.anchor.paymentsservice.circle;
-
-import lombok.Data;
-
-@Data
-public class CircleWalletResponse {
-    CircleWallet data;
-}
