+++ src/main/java/org/stellar/anchor/paymentsservice/circle/model/response/CircleError.java
@@ -1,4 +1,4 @@
-package org.stellar.anchor.paymentsservice.circle;
+package org.stellar.anchor.paymentsservice.circle.model.response;
 
 import lombok.Data;
